requirements.txt
Keras==2.4.3
numpy==1.19.5
pandas==1.2.4
Pillow==8.1.2
scipy==1.6.3
streamlit==0.80.0
tensorflow==2.5.0rc1
toml==0.10.2